refactor: клавиатуры разнесены по модулям
