+++ handlers/set_user_name.py
@@ -3,7 +3,8 @@ from aiogram.filters import StateFilter
 from aiogram.types import Message, CallbackQuery
 from lexicon.lexicon_ru import LEXICON_SET_USER_NAME, LEXICON_MENU_BUTTONS
 from database.database import users_db
-from keyboards.keyboards import menu_kb, yes_no_name_kb
+from keyboards.menu_kb import menu_kb
+from keyboards.user_name_setting import yes_no_name_kb
 
 from aiogram.fsm.context import FSMContext
 from fsm.fsm import FSM_SET_NAME
