+++ text_creator.py
